test: fix screen test by waiting for elements
Resolves: DCMAW-8868
